requirements.txt
aiogram==3.21.0
python-dotenv
aiofiles
Kurigram==2.2.7
TgCrypto
aiohttp
aiohttp-socks
redis